Рефактор tsv-file-reader.ts и добавление потоков
