+++ package.json
@@ -30,6 +30,7 @@
     "eslint-config-htmlacademy": "9.1.1",
     "rimraf": "5.0.1",
     "ts-node": "10.9.1",
+    "tslib": "2.6.2",
     "typescript": "5.2.2",
     "json-server": "0.17.3"
   },
