+++ src/shared/libs/file-reader/index.ts
@@ -1,2 +1,2 @@
 export * from './file-reader.interface.js';
-export * from './tsv-file-reader.js';
+export * from './tsv-file-reader-old.js';
